Add dimension and biome NBT serialization

+++ src/Redstone.Common/Codecs/Biomes/BiomeMoodSound.cs
@@ -5,9 +5,9 @@ namespace Redstone.Common.Codecs.Biomes
     public class BiomeMoodSound
     {
         [JsonPropertyName("tick_delay")]
-        public uint TickDelay { get; set; }
+        public int TickDelay { get; set; }
 
-        public uint Offset { get; set; }
+        public double Offset { get; set; }
 
         public string Sound { get; set; }
